Implement recycle protocol
By recycling a channel, a new funding transaction is pushed on-chain,
spending the previous funding transaction.

In this state, the protocol has little use but the next step would be to
add custom inputs and outputs to this `recycle` transaction to allow
recharging or withdrawing bitcoin from the channel while only spending
the fees of one transaction.

This will be probably be renamed to `rebalance` or `splicing` once said
features are done.