Subtract all fees when building split transaction
Instead of passing in an amount per output that already takes into
account the fees used in `TX_c`, we pass in the raw balances and
subtract the fees for both `TX_c` and `TX_s` in the constructor for
`TX_s`.

Also, simplify the test `e2e_force_close_after_updates`, since the
test was already failing after one update.