Attempt uniform identifier usage
Currently we have `foo_alice`, `alice_foo` and spread out all over the
place. This is problematic for two reasons, firstly it is non-uniform
so one cannot guess an identifier form but has to check it. We should
be uniform. Secondly using 'alice' and 'bob' is way to verbose, it
clutters the code making it harder to read what the actual variable
is. Using `a_foo` and `b_foo` increases clarity with no loss of
meaning.

This PR leaves 'balance' identifiers still non-uniform, there are a
whole bunch of these, many with other sub-parts. Fix for these to
follow.