Use lowercase tx instead of TX
`TX` was (presumably) chosen to match the whitepaper. The whitepaper
and code are different forms, using the identifier forms from the
whitepaper makes the code harder to read and reason about for the
following reasons

- It messes with the tooling (syntax highlighting is all wrong)
- It is non-idiomatic so the readers eye is constantly drawn to them
- It is (arguably) harder to type

On top of this, since the `TX` is always used with another separater
it adds no additional information e.g., `tx_s` and `TX_s` give the
same amount of information just as clearly as each other.

Favour idiomatic rust over the whitepaper, use lowercase 'tx'.