Make serde cfg_attr less noisy
We currently have a bunch of places in the codebase that use the `cfg_attr` to
control serialization. We use the fully qualified path of `as_sat()` which often
makes the attribute format over 4 lines. If we import the `as_sat` function then
we can remove the fully qualified path and reduce the attribute to a single
line. This makes the code way less noisy and much more gentle on the eyes of the
reader.