Add step macros
Currently we have loads of duplicate code for sending/receiving messages and
interpreting the received message.

Add two declarative macros that combine these three steps, `step` and
`step_wallet`. The second takes an additional `wallet` parameter.

We return the transport from the macros, it would be nice to not have to do this
but I can't seem to figure that out.