Create close transaction while creating close::State0
We clone all the channel data during construction of `State0`. In `compose` and
`interpret` we create the close transaction based on this sate data. This call
can fail but we now already if it will fail when we create the `State0` object.
Instead of waiting until compose/interpret just check the balance is sufficient
during construction and return the error then. This has the added advantage
bringing the `close` impls of compose/interpret inline with the impls in the
other protocols.